requirements.txt

streamlit
python-docx
docx2txt
PyPDF2
graphviz
docxtpl
